Organized components into folders for easier reading
