+++ src/components/index.js
@@ -1,5 +1,5 @@
 /* All exports */
 
 export { default as App } from "./App";
-export { default as BookmarkCard } from "./BookmarkCard.js";
-export { default as SearchBar } from "./SearchBar.js";
+export { default as BookmarkCard } from "./BookmarkCard/BookmarkCard.js";
+export { default as SearchBar } from "./SearchBar/SearchBar.js";
